add gun B and guid page and audio

+++ projectIO/src/main/java/com/example/paintio/Position.java
@@ -2,10 +2,10 @@
 package com.example.paintio;
 
 public class Position implements Comparable<Position> {
-    public double x;
-    public double y;
+    public int x;
+    public int y;
 
-    public Position(double x, double y) {
+    public Position(int x, int y) {
         this.x = x;
         this.y = y;
     }
